+++ mono_cli/lib/src/cli/parser_args.dart
@@ -28,6 +28,7 @@ class ArgsCliParser implements CliParser {
       ..addOption('concurrency', abbr: 'j')
       ..addOption('order')
       ..addFlag('dry-run', defaultsTo: false)
+      ..addFlag('check', defaultsTo: false)
       ..addMultiOption('targets', abbr: 't');
 
     final results = parser.parse(rest);
@@ -48,6 +49,7 @@ class ArgsCliParser implements CliParser {
     put('concurrency', results['concurrency']);
     put('order', results['order']);
     if (results['dry-run'] == true) put('dry-run', 'true');
+    if (results['check'] == true) put('check', 'true');
     put('targets', results['targets']);
 
     final targets = <TargetExpr>[];